id int to long

+++ src/main/java/org/code_factory/jpa/nestedset/JpaNode.java
@@ -60,7 +60,7 @@ class JpaNode<T extends NodeInfo> implements Node<T> {
         this.type = (Class<T>) node.getClass();
     }
 
-    @Override public int getId() {
+    @Override public Long getId() {
         return this.node.getId();
     }
